Changed numerical values from scale to spinbutton to fix issue #67
In newer Cinnamon updates, the scale component is not displaying numbers
anymore, which makes it hard to pick the desired number of minutes we
want for a Pomodoro session for example.

This change replaces all "scale" components by "spinbutton" components
for selecting specific numbers (minutes or number of pomodori in a
session).